Renamed TipShaperTest.py TipShaper.py and added TipShaperTest.py
